+++ pandas/tests/frame/test_block_internals.py
@@ -254,15 +254,6 @@ class TestDataFrameBlockInternals:
         if not compat.is_platform_windows():
             f("M8[ns]")
 
-    def test_copy(self, float_frame, float_string_frame):
-        cop = float_frame.copy()
-        cop["E"] = cop["A"]
-        assert "E" not in float_frame
-
-        # copy objects
-        copy = float_string_frame.copy()
-        assert copy._mgr is not float_string_frame._mgr
-
     def test_pickle(self, float_string_frame, timezone_frame):
         empty_frame = DataFrame()
